Added PU handling
For MC the PU reweighing for runs C,D,E and F is hardcoded in the nTupler. Added *puHCalc.py* for calculating the values as described in the PU reweighing twiki. Also added script *getJSONforPUrange.py* to modify JSON file for data so it contains only a predefined PU range.